requirements.txt
fastapi
uvicorn
python-binance
pydantic
